in the middle of changing table to angular material

+++ Template.Domain/Entities/ModuleProfile.cs
@@ -1,11 +1,11 @@
 ﻿namespace Template.Domain.Entities
 {
-	public class ModuleProfile
-	{
-		public int ModuleId { get; set; }
-		public int ProfileId { get; set; }
+    public class ModuleProfile
+    {
+	public long ModuleId { get; set; }
+	public long ProfileId { get; set; }
 
-		public virtual Module Module { get; set; }
-		public virtual Profile Profile { get; set; }
-	}
+	public virtual Module Module { get; set; }
+	public virtual Profile Profile { get; set; }
+    }
 }
