+++ Template.Domain/Entities/User.cs
@@ -8,7 +8,7 @@
         public string Code { get; set; }
         public bool IsAuthorised { get; set; }
 
-        public int ProfileId { get; set; }
+        public long ProfileId { get; set; }
         public virtual Profile Profile { get; set; }
     }
 }
